Cleanup of ResolvedPoint generic type by adding tags and defining interface.
Former-commit-id: aa8822749dd9976fe88ce1750be3d3a358654bee

+++ Core/Osm.Map/Layers/Custom/CustomLayer.cs
@@ -378,7 +378,7 @@ namespace Osm.Map.Layers.Custom
 
         public ElementText AddText(
             int color,
-            int size,
+            float size,
             string text,
             GeoCoordinate coordinate)
         {
